pyproject.toml
[project]
name = "agent-definitions"
version = "0.1.0"
description = "Botburrow agent configuration repository"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20.0",
    "boto3>=1.34.0",
    "httpx>=0.27.0",
]

# Scripts are run directly: python scripts/validate.py
# Not installed as package entry points

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = []  # Scripts-only repo, no Python packages to install

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
